refactor: move main into app package
